\name{testpaket-package}
\alias{testpaket-package}
\alias{testpaket}
\docType{package}
\title{
\packageTitle{testpaket}
}
\description{
\packageDescription{testpaket}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testpaket}
\packageIndices{testpaket}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{testpaket}

Maintainer: \packageMaintainer{testpaket}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
